Añadir los productos al carrito de compras y mostrarlos en la vista carrito

+++ src/main/java/com/pormarweb/ecommerce/model/DetalleOrden.java
@@ -16,7 +16,7 @@ public class DetalleOrden {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
 	private String nombre;
-	private Double cantidad;
+	private Integer cantidad;
 	private Double precio;
 	private Double total;
 	
@@ -30,7 +30,7 @@ public class DetalleOrden {
 	public DetalleOrden() {
 	}
 	
-	public DetalleOrden(Integer id, String nombre, Double cantidad, Double precio, Double total) {
+	public DetalleOrden(Integer id, String nombre, Integer cantidad, Double precio, Double total) {
 		super();
 		this.id = id;
 		this.nombre = nombre;
@@ -54,11 +54,11 @@ public class DetalleOrden {
 		this.nombre = nombre;
 	}
 	
-	public Double getCantidad() {
+	public Integer getCantidad() {
 		return cantidad;
 	}
 	
-	public void setCantidad(Double cantidad) {
+	public void setCantidad(Integer cantidad) {
 		this.cantidad = cantidad;
 	}
